moved Tests in new project ObjectPrinting.Tests
